[FIX] hr_payroll: salary slip refund without recompute of lines
Before this commit, when a refund of a salary slip was done, the 'Salary
Computation' lines where recompute from the 'Worked Days & Inputs'
lines. This generates falsely data if the structure or the contract
changes between the salary slip and the refund.

Now, when a refund is done, all the lines are copied from the salary
slip to be coherent with the salary slip.

opw-2081454

closes odoo/odoo#39412

Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>