[FIX] hr_payroll: fix traceback on payroll creation
Have a [DEMO] employee with an active (i.e. open) contract
Encode an holiday (i.e 15->16 April)
Encode a time off for [DEMO] which overlaps the holiday (i.e. 13->17
April)
Try to generate a payroll for the employee

Traceback will popup, blocking the auto generation of the "worked days"
lines. This happens because the holiday overlaps causes a recordset
filled with multiple values to be queried for the id. Taking just the
first one should be ok since the period taken into account is the full
one (merged overlaps).

opw-2236420

closes odoo/odoo#49598

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>